@@ -4,7 +4,7 @@ package com.thunderduck.types;
  * Data type representing a single-precision 32-bit floating point number.
  * Maps to DuckDB FLOAT and Spark FloatType.
  */
-public class FloatType extends DataType {
+public final class FloatType implements DataType {
 
     private static final FloatType INSTANCE = new FloatType();
 
@@ -32,5 +32,10 @@ public class FloatType extends DataType {
     @Override
     public int hashCode() {
         return typeName().hashCode();
+    }
+
+    @Override
+    public String toString() {
+        return typeName();
     }
 }
